% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCite.R
\name{addCite}
\alias{addCite}
\title{include citation for package in markdown}
\usage{
addCite(bib = NULL, packages = NULL, type = "T")
}
\arguments{
\item{bib}{bibliography for RefManageR otherwise put bibtex citation in}

\item{packages}{vector of package names}

\item{type}{T for in text A for other}
}
\description{
include citation for package in markdown
}
