% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignBWA.R
\name{alignBWA}
\alias{alignBWA}
\title{Simple subread aligner}
\usage{
alignBWA(fileName, projName, runId,
  outRoot = file.path("/scratch/pschofield/Projects", projName),
  outName = "Data/alignment/bwa", noSub = F, scpIt = T,
  bwaMod = "apps/bwa/0.7.13/gcc-4.4.7",
  samtoolsMod = "apps/samtools/0.1.19/gcc/4.4.7", mem = "32Gb",
  refGenome = NULL, ncores = 8, fqExt = ".fastq", pe = T)
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
This is a simpler subread aligner that runs of single fastq files rather than individual lane fastqs
from the HiSeq. There maybe QC advantages of aligning the individual fastqc but sometimes it is just
simpler to merge before the event.
}
