% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignStar.R
\name{alignStar}
\alias{alignStar}
\title{Simple STAR aligner does everything counts junctions and genes}
\usage{
alignStar(projName, sampleID, inputPath,
  outRoot = file.path("/scratch/pschofield/Projects", projName),
  outName = "Data/alignment/star", noSub = F, scpIt = T,
  starMod = "apps/star/2.5.1b/gcc-5.1.0",
  samtoolsMod = "apps/samtools/0.1.19/gcc/4.4.7", refSequence = NULL,
  refGenome = NULL, ncores = 16, fqExt = ".fastq", pe = T)
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
This is a simpler STAR aligner that runs of single fastq files rather than individual lane fastqs from
the HiSeq. There maybe QC advantages of aligning the individual fastqc but sometimes it is just
simpler to merge before the event.
}

