% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSubjunc.R
\name{alignSubjunc}
\alias{alignSubjunc}
\title{Simple subjunc aligner does everything counts junctions and genes}
\usage{
alignSubjunc(projName, sampleID, inputPath, outStub = NULL,
  outRoot = file.path("/scratch/pschofield/Projects", projName),
  outName = "Data/alignment/star", noSub = F, scpIt = T,
  subreadMod = "apps/subread/1.5.0-p3/gcc-4.4.7",
  samtoolsMod = "apps/samtools/1.2/gcc/4.4.7", mem = "32Gb",
  stranded = "fr", nmap = "2", refGenome = NULL, ncores = 16,
  fqExt = ".fastq", pe = T)
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
This is a subjunc aligner that runs of single fastq files rather than individual lane fastqs from
the HiSeq. There maybe QC advantages of aligning the individual fastqc but sometimes it is just
simpler to merge before the event.
}
