% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQDNAFilter.R
\name{applyQDNAFilter}
\alias{applyQDNAFilter}
\title{Apply QDNAseq filter}
\usage{
applyQDNAFilter(rd, filterGR = filter150kb, outfile = NULL)
}
\arguments{
\item{rd}{rangeddata input object}

\item{outfile}{a file to write the output to a bigwig file for viewing in IGV}

\item{filter}{a filter in the form of a granges object}
}
\description{
function to apply a filter to a HMMcopy RangedData object and return a filtered RangedData object
}
