% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCommit.R
\name{autoCommit}
\alias{autoCommit}
\title{Loop through Project code committing and pulling as necessary}
\usage{
autoCommit(
  projDir = file.path(Sys.getenv("HOME"), "repositories"),
  commitMessage = "Automated ",
  oncampus = F,
  mws = F
)
}
\arguments{
\item{projDir}{Project Root}

\item{pat}{pattern for directory names}

\item{subdir}{names of subdirectories that include code}
}
\description{
currently my build of libgit is not coping with HTTPs protocol need to fix this at some point
until then dropping to system shell to push and pull
}
