% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToBed.R
\name{bamToBed}
\alias{bamToBed}
\title{create bed file from a bam file}
\usage{
bamToBed(inFile, outDir, projName, paired = "",
  outRoot = file.path("/scratch/pschofield/Projects", projName),
  bedtoolsMod = "apps/bedtools/2.25.0/gcc-4.4.7")
}
\arguments{
\item{inFile}{input bam file}

\item{outDir}{output directory}

\item{projName}{name of the project}
}
\description{
create bed file from a bam file
}
