% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base2dec.R
\name{base2dec}
\alias{base2dec}
\title{Convert a bit string to a number}
\usage{
base2dec(base_number, base = 2)
}
\arguments{
\item{base_number}{(the number to convert)}

\item{base}{the base to convert to}
}
\description{
Convert a bit string to a number
}
