% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigcor.R
\name{bigcor}
\alias{bigcor}
\title{large correlation matrix with ff}
\usage{
bigcor(x, nblocks = 10, verbose = TRUE, absolute = F)
}
\arguments{
\item{x}{the matrix}

\item{nblocks}{required to be factor of ncols x}

\item{verbose}{report whats going on}
}
\description{
large correlation matrix with ff
}

