% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{build_master_list}
\alias{build_master_list}
\title{build master list}
\usage{
build_master_list(
  repoNames,
  git_directory = file.path(Sys.getenv("HOME"), "Repos"),
  recur = FALSE,
  htmlroot = "uol",
  html_directory = file.path("/srv", "http"),
  pub = F
)
}
\description{
build master list
}
