% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CitePres.R
\name{CitePres}
\alias{CitePres}
\title{Presentation citation}
\usage{
CitePres(bib = NULL, ckey = "", alim = 2)
}
\arguments{
\item{alim}{maximum number of authors before et. al.}

\item{biblio}{bibliography}

\item{key}{bibtex citation key}
}
\description{
A funtion to add citations in presentations without having to have a bibliography at the end
}
