% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAnnoGenes.R
\name{cnvAnnoGenes}
\alias{cnvAnnoGenes}
\title{annotate CNV segmentation and copynumber call}
\usage{
cnvAnnoGenes(segData, annoRdata, pl)
}
\arguments{
\item{segData}{segmentation dataset}

\item{annoRdata}{gene Annotation file}
}
\description{
annotate CNV segmentation and copynumber call
}
