% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvAnnoSegs.R
\name{cnvAnnoSegs}
\alias{cnvAnnoSegs}
\title{annotate CNV segmentation and copynumber call}
\usage{
cnvAnnoSegs(segData, annoRdata, cytoFile, pl)
}
\arguments{
\item{segData}{segmentation dataset}

\item{cytoFile}{cytoband file}
}
\description{
annotate CNV segmentation and copynumber call
}
