% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCopyNumber.R
\name{cnvCopyNumber}
\alias{cnvCopyNumber}
\title{Function for reading in wig file from a single sample}
\usage{
cnvCopyNumber(readcountFile, gcFile, mapFile, myChromosome = c(1:22, "X",
  "Y"))
}
\description{
Function for reading in wig file from a single sample
}
