% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvCalcLogRatio.R
\name{cnvLogRatio}
\alias{cnvLogRatio}
\title{Function for reading in wig file from a pair of samples}
\usage{
cnvLogRatio(tumourFile, normalFile, gcFile, mapFile,
  myChromosome = c(1:22, "X", "Y"))
}
\description{
Function for reading in wig file from a pair of samples
}
