% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvSegment.R
\name{cnvSegment}
\alias{cnvSegment}
\title{segment as copy number data set}
\usage{
cnvSegment(ds, pl, mkwigs = T)
}
\arguments{
\item{ds}{copynumber dataset}
}
\description{
segment as copy number data set
}
