% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collateRPG.R
\name{collateRPG}
\alias{collateRPG}
\title{collate a directory of STAR .ReadsPerGene.out.tab files}
\usage{
collateRPG(directoryName, locDir = NULL, strandedness = "US", force = F)
}
\arguments{
\item{dirname}{directory name}
}
\description{
collate a directory of STAR .ReadsPerGene.out.tab files
}

