% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compilequarto.R
\name{compile_quarto}
\alias{compile_quarto}
\title{Render the current Quarto document}
\usage{
compile_quarto(
  fileName = .fileName,
  projName = .projName,
  gitRepo = "uol",
  sysRoot = Sys.getenv("HOME"),
  user = Sys.getenv("USER"),
  outPath = file.path(Sys.getenv("HOME"), "Projects"),
  nbPath = file.path("/srv", "http"),
  codePath = file.path(sysRoot, "Repos", gitRepo),
  docPath = file.path(sysRoot, "Projects"),
  silent = FALSE,
  setHome = FALSE,
  toPDF = FALSE,
  toDOCX = FALSE,
  toHTML = TRUE,
  ext = "qmd",
  dbg = FALSE
)
}
\arguments{
\item{fileName}{file name (without extension)}

\item{projName}{project name}

\item{gitRepo}{Git repository (default "uol")}

\item{sysRoot}{home directory}

\item{user}{User id}

\item{outPath}{local directory where rendered files will be written}

\item{nbPath}{location for HTML output intended for server use}

\item{codePath}{path to source code (the Quarto document)}

\item{docPath}{(unused) document tree path}

\item{silent}{if TRUE, do not automatically display the output URL}

\item{setHome}{if TRUE, the rendered HTML file will be named "index.html"}

\item{toPDF}{if TRUE, render PDF output}

\item{toDOCX}{if TRUE, render DOCX output}

\item{toHTML}{if TRUE, render HTML output}

\item{ext}{file extension, default is "qmd" for Quarto documents}

\item{dbg}{debug flag (unused)}

\item{htmlUP}{if TRUE, push the HTML file to the nbPath location (if available)}
}
\description{
This function renders a Quarto document and saves the output in a local location.
}
