% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{create_meta_yaml}
\alias{create_meta_yaml}
\title{create project meta data}
\usage{
create_meta_yaml(projDir, projName, projDesc, owner, contact, overwrite = F)
}
\description{
create a YAML document with meta data about the project for indexing
}
