% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRUKlib.R
\docType{package}
\name{CRUKlib}
\alias{CRUKlib}
\alias{CRUKlib-package}
\title{CRUKlib}
\description{
This is a library of functions I have written to ease my work at CRUK Manchester Institute 
It comprises of various functions for accessing the job queuing system of troodon and submitting 
jobs. I also has some plot functions for standard plots. It will probably integrate with the 
snakemake pipeline system when that comes on stream
}

