% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custHMMseqment.R
\name{custHMMsegment}
\alias{custHMMsegment}
\title{Custom CNV segmentation}
\usage{
custHMMsegment(cn, pl, outFile = NULL)
}
\arguments{
\item{cn}{corrected copy number RangedData object}

\item{outFile}{if specified writes the segmentation out as a wig file}

\item{segmentationMu}{comma separated string of values}

\item{long}{if true performs adjustment to favour long segments}
}
\description{
uses HMMcopy HMMsegment function but applys a custom segmentation mu string
}
