% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupCNV.R
\name{dedupCNV}
\alias{dedupCNV}
\title{run the picard and GATK pre processing steps on bam alignments}
\usage{
dedupCNV(inBAM, projName, refGenome, gatkStep = T,
  tmpDir = file.path("/scratch/pschofield/Projects", projName, "tmp"),
  picardDir = "/apps/modules/pkg/apps/picardtools/1.96/noarch",
  gatkDir = "/apps/modules/pkg/apps/gatk/3.1.1/noarch",
  javaMod = "compilers/java/1.7.0_80/noarch",
  gatkMod = "apps/gatk/3.1.1/noarch",
  picardMod = "apps/picardtools/1.96/noarch",
  samtoolsMod = "apps/samtools/0.1.19/gcc/4.4.7",
  gatkRes = "/data/compbio/hsleong/Genome/gatk/hg19/bundle", ncores = 1,
  scpIt = T, noSub = F)
}
\arguments{
\item{filename}{merged bam file}
}
\description{
run the picard and GATK pre processing steps on bam alignments
}
