% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delWeek.R
\name{delWeek}
\alias{delWeek}
\title{getWeek return the data for the week comencing}
\usage{
delWeek(startDate, projDir = "/Users/pschofield/Projects/time_sheet/",
  dbName = "timeSheet.db")
}
\arguments{
\item{start}{date}
}
\description{
get the time_sheet data for a particular week
}

