% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_gitbook.R
\name{deploy_gitbook}
\alias{deploy_gitbook}
\title{Deploy a GitBook to a Web Server}
\usage{
deploy_gitbook(
  book_path = ".",
  book_name = NULL,
  server_user = "pietas",
  server_host = "localhost",
  server_path = file.path("srv", "http", "uol")
)
}
\arguments{
\item{book_path}{Character. The local path to the GitBook source directory. Defaults to `"."` (current
directory).}

\item{book_name}{Character. The name of the book directory}

\item{server_user}{Character. The SSH username for the remote server. Defaults to
`"user"`.}

\item{server_host}{Character. The hostname or IP address of the remote server. Defaults to
`"yourserver.com"`.}

\item{server_path}{Character. The destination path on the web server where the
book should be deployed.}
}
\value{
No return value. This function performs rendering and deployment as a side effect.
}
\description{
This function wraps `bookdown::render_book()` to render a GitBook and 
then automatically copies necessary assets (such as images and HTML pages) 
to the correct location on a web server.
}
