% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{discover_repos}
\alias{discover_repos}
\title{discover top-level repos/projects automatically}
\usage{
discover_repos(
  git_directory = file.path(Sys.getenv("HOME"), "repositories"),
  exclude = c(".git", ".Rproj.user", "pers", "archive", "protocols", "public",
    "public_pages", "dotfiles")
)
}
\arguments{
\item{exclude}{character vector of folder names to skip (case-insensitive)}
}
\description{
discover top-level repos/projects automatically
}
