% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.R
\name{disp_tab}
\alias{disp_tab}
\title{New display data}
\usage{
disp_tab(
  dataset,
  number = NULL,
  disp = TRUE,
  limited = FALSE,
  buttons = FALSE,
  plen = NULL,
  caption = NULL,
  fixh = TRUE,
  fixc = list(leftColumns = 1),
  sigf = 3,
  wrap_columns = NULL
)
}
\description{
New display data
}
