% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{display_project_index}
\alias{display_project_index}
\title{Display Project Index}
\usage{
display_project_index(projName, gitRepo = .gitRepo, htmlroot = htmlroot)
}
\description{
Display Project Index
}
