% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayURL.R
\name{displayURL}
\alias{displayURL}
\title{display a web page nicely in chrome}
\usage{
displayURL(urlpath = "http://localhost", browser_path = "/usr/bin/brave")
}
\description{
display a web page nicely in chrome
}
