% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filter150kb}
\alias{filter150kb}
\title{filter150kb}
\format{a GRanges object}
\usage{
filter150kb
}
\description{
150kilobase window filter for HMMcopy data generated using QDNAseq from 38 1000 genome project
datasets
}
\keyword{datasets}
