% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCNValignmentStats.R
\name{genCNValignmentStats}
\alias{genCNValignmentStats}
\title{generate alignments stats for BWA CNV alignments}
\usage{
genCNValignmentStats(sid, bamDir, fqsDir, projName,
  remRoot = file.path("/scratch/pschofield/Projects", projName),
  outDir = "Analysis/mapping_stats", noSub = F, ncores = 1,
  scpIt = T, mem = "32Gb",
  samtoolsMod = "apps/samtools/0.1.19/gcc/4.4.7", ufExt = ".bam",
  mdExt = "_dedup.bam", rdExt = "_rmdup.bam")
}
\arguments{
\item{bamDir}{directory of alignment files}
}
\description{
generate alignments stats for BWA CNV alignments
}
