% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCNVmapGC.R
\name{genCNVmapGC}
\alias{genCNVmapGC}
\title{generate wiggle tracks for HMMcopy}
\usage{
genCNVmapGC(projName, remRoot = "/scratch/pschofield/Projects",
  outPath = "Analysis/wigs", noSub = F, ncores = 1, scpIt = T,
  mem = "16Gb", windowSize, chrNames = c(seq(1, 22), "X", "Y"), genomeFasta,
  mapFile, hmmCopyMod = "apps/hmmcopy/0.99.0/gcc-4.4.7")
}
\arguments{
\item{projName}{the project}

\item{sampleID}{the bam files}

\item{location}{of bam files}
}
\description{
generate wiggle tracks for HMMcopy
}
