% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCNVwigs.R
\name{genCNVwigs}
\alias{genCNVwigs}
\title{generate wiggle tracks for HMMcopy}
\usage{
genCNVwigs(sampleID, projName, bamDir, bamExt = "_dedup_realn_recal.bam",
  outDir = file.path("/scratch/pschofield/Projects", projName,
  "Analysis/wigs"), noSub = T, ncores = 1, scpIt = T, mem = "16Gb",
  windowSize, chrNames = c(seq(1, 22), "X", "Y"),
  hmmCopyMod = "apps/hmmcopy/0.99.0/gcc-4.4.7")
}
\arguments{
\item{sampleID}{the bam files}

\item{projName}{the project}

\item{location}{of bam files}
}
\description{
generate wiggle tracks for HMMcopy
}
