% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCovTracks.R
\name{genCovTracks}
\alias{genCovTracks}
\title{generate some bigwig tracks from paired reads}
\usage{
genCovTracks(projName, fileList, outPath, paired = T)
}
\arguments{
\item{projName}{project name}

\item{fileList}{bam files to use}

\item{outPath}{directory to write output to}
}
\description{
generate some bigwig tracks from paired reads
}
