% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHeaders.R
\name{generateHeaders}
\alias{generateHeaders}
\title{Some R functions for processing and analysing of genomics data}
\usage{
generateHeaders(PBSscript, jobName = "myrun",
  jobDescription = "My analysis", Nnodes = 1, Nproc = 16,
  Memory = "32gb", Walltime = "30:00:00", delayTil = NULL,
  environ = TRUE, email = "pieta.schofield@cruk.manchester.ac.uk",
  overwrite = FALSE, log.dir = "/lustre/scratch/pschofield/tmp")
}
\arguments{
\item{PBSscript}{name of the script file}

\item{jobName}{name of job}

\item{jobDescription}{description of job}

\item{Nnodes}{number of nodes requested}

\item{Nproc}{number of cores requested}

\item{Memory}{size of memory requested}

\item{Walltime}{limit to run time}

\item{delayTil}{time delay for start}

\item{environ}{pass current environment variables}

\item{email}{address to notify of job progress}

\item{overwrite}{boolean overwrite existing scripts}

\item{log.dir}{directory output logs}
}
\description{
Generate headers for running batch PBS jobs
}
