% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genPBSHead.R
\name{genPBSHead}
\alias{genPBSHead}
\title{Batch job script creation}
\usage{
genPBSHead(
  jobName = "myrun",
  jobDescription = "My analysis",
  Nnodes = 1,
  Nproc = 16,
  Memory = "32gb",
  Walltime = "1:00:00",
  delayTil = NULL,
  environ = TRUE,
  email = "pieta.schofield@manchester.ac.uk",
  overwrite = FALSE,
  emailMode = "a",
  log.dir = "/data/compbio/pschofield/logs",
  dirSet = NULL,
  jobDep = NULL
)
}
\arguments{
\item{jobName}{name of job}

\item{jobDescription}{description of job}

\item{Nnodes}{number of nodes requested}

\item{Nproc}{number of cores requested}

\item{Memory}{size of memory requested}

\item{Walltime}{limit to run time}

\item{delayTil}{time delay for start}

\item{environ}{pass current environment variables}

\item{email}{address to notify of job progress}

\item{overwrite}{boolean overwrite existing scripts}

\item{log.dir}{directory output logs}
}
\description{
Generate headers for running batch PBS jobs
}
