% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSGEHead.R
\name{genSGEHead}
\alias{genSGEHead}
\title{Batch job script creation}
\usage{
genSGEHead(
  jobName = "myrun",
  jobDescription = "My analysis",
  Nnodes = 1,
  Nproc = 16,
  Memory = NULL,
  email = "pieta.schofield@manchester.ac.uk",
  emailMode = "ae",
  logDir = "/scratch/log",
  pe = "smp.pe"
)
}
\arguments{
\item{jobName}{name of job}

\item{jobDescription}{description of job}

\item{Nnodes}{number of nodes requested}

\item{Nproc}{number of cores requested}

\item{Memory}{size of memory requested}

\item{email}{address to notify of job progress}

\item{Walltime}{limit to run time}

\item{delayTil}{time delay for start}

\item{environ}{pass current environment variables}

\item{overwrite}{boolean overwrite existing scripts}

\item{log.dir}{directory output logs}
}
\description{
Generate headers for running batch PBS jobs
}
