% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSlurmHead.R
\name{genSlurmHead}
\alias{genSlurmHead}
\title{Batch job script creation}
\usage{
genSlurmHead(jobName = "myrun", jobDescription = "My analysis",
  Nnodes = 1, Nproc = 16, Memory = "32gb", Walltime = "24:00:00",
  email = "pieta.schofield@cruk.manchester.ac.uk", emailMode = "ALL",
  logDir = "/home/camp/schofip/logs", Partition = "compute")
}
\arguments{
\item{jobName}{name of job}

\item{jobDescription}{description of job}

\item{Nnodes}{number of nodes requested}

\item{Nproc}{number of cores requested}

\item{Memory}{size of memory requested}

\item{Walltime}{limit to run time}

\item{email}{address to notify of job progress}

\item{logDir}{directory output logs}
}
\description{
Generate headers for running batch PBS jobs
}
