% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSRstats.R
\name{genSRstats}
\alias{genSRstats}
\title{generate stats files for subread / subjunc alignments for summarisation}
\usage{
genSRstats(fileName, projName, ncores = 2, noSub = F, mem = "8Gb",
  scpIt = T, samtoolsMod = "apps/samtools/1.3.1/gcc-4.4.7")
}
\arguments{
\item{fileName}{bam file}

\item{projName}{}
}
\description{
generate stats files for subread / subjunc alignments for summarisation
}
