% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcreds.R
\name{getcreds}
\alias{getcreds}
\title{open a database using the credentials file}
\usage{
getcreds(
  credir = file.path(Sys.getenv("HOME"), ".config"),
  credfile = file.path(credir, "psqlcred.yml"),
  credname = NULL
)
}
\description{
open a database using the credentials file
}
