% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProject.R
\name{getProject}
\alias{getProject}
\title{get information from project xlsx}
\usage{
getProject(filename, sumSheet = "Summary", fileSheet = "Files", ssCol = 1,
  seCol = 3, ssRow = 1, fsCol = 1, feCol = 6, fsRow = 1)
}
\arguments{
\item{filename}{name of the project file}

\item{sumSheet}{name of summary sheet}

\item{fileSheet}{name of raw data files sheet}

\item{ssCol}{start column summary sheet}

\item{seCol}{end column summary sheet}

\item{ssRow}{start row summary sheet}

\item{fsCol}{start column file sheet}

\item{feCol}{end column file sheet}

\item{fsRow}{start row file sheet}
}
\description{
get information from project xlsx
}

