% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpikeInNorm.R
\name{getSpikeInNorm}
\alias{getSpikeInNorm}
\title{get results from nicks spike-in db}
\usage{
getSpikeInNorm(expID = NULL, custom = NULL)
}
\arguments{
\item{custom}{custom sql string}

\item{expid}{}
}
\description{
get results from nicks spike-in db
}
