% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeek.R
\name{getWeek}
\alias{getWeek}
\title{getWeek return the data for the week comencing}
\usage{
getWeek(startDate, endDate,
  projDir = "/Users/pschofield/Projects/time_sheet/",
  dbName = "timeSheet.db")
}
\arguments{
\item{start}{date}
}
\description{
get the time_sheet data for a particular week
}

