% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{go}
\alias{go}
\title{edit most recent note}
\usage{
go(
  pname = "notes",
  fname = NULL,
  tname = "git_tree",
  rname = "/Users/pschofield/",
  oflag = TRUE
)
}
\arguments{
\item{pname}{project name}

\item{fname}{file name}

\item{tname}{tree name}

\item{rname}{root of tree}

\item{oflag}{open a file}
}
\description{
edit most recent note
}
