% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmAnalysis.R
\name{hmAnalysis}
\alias{hmAnalysis}
\title{use heatmap and tree cutting}
\usage{
hmAnalysis(exprMat, rowV = NA, colV = NA, ncuts = 6, filename = NA,
  mainTitle = "Heatmap", cr = 1.5, cc = 0.5)
}
\description{
use heatmap and tree cutting
}
