% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_site.R
\name{index_site}
\alias{index_site}
\title{index my site}
\usage{
index_site(
  html_dir = file.path("/srv", "http"),
  index_file = file.path(html_dir, "search_index.json"),
  db = F
)
}
\description{
index the website
}
