% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_only.R
\name{interactive_only}
\alias{interactive_only}
\title{Run only in interactive sessions}
\usage{
interactive_only(expr)
}
\arguments{
\item{expr}{Code to evaluate (not quoted)}
}
\value{
NULL invisibly
}
\description{
Run only in interactive sessions
}
