% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lck.R, R/listcitekeys.R
\name{lck}
\alias{lck}
\title{list the citation keys for a package}
\usage{
lck(packages, title = F)

lck(packages, title = F)
}
\arguments{
\item{the}{package name}
}
\description{
list the citation keys for a package

list the citation keys for a package
}
