% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll.R
\name{ll}
\alias{ll}
\title{ll list full info on objects}
\usage{
ll(name = NULL, pos = 1L, envir = as.environment(pos), allNames = FALSE,
  pattern = ".*", order.by = NULL, decreasing = FALSE, head = FALSE,
  n = 100)
}
\arguments{
\item{name}{see ls()}

\item{pos}{see ls()}

\item{envir}{see ls()}

\item{allNames}{see ls()}

\item{pattern}{see ls()}

\item{decreasing}{reverse direction}

\item{head}{restrict number}

\item{n}{number to list}

\item{orderBy}{field to orderby}
}
\description{
This is a steal however it probably needs a rewrite in the plyr world
}
