% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSubreadOutput.R
\name{loadSubreadOutput}
\alias{loadSubreadOutput}
\title{Load Fcount output from : DEXSeq_after_Fcount.R into DEXSeq
 Copyright 2015 Vivek Bhardwaj (bhardwaj@ie-freiburg.mpg.de). Licence: GPLv3.
 Read Fcount output and convert to dxd}
\usage{
loadSubreadOutput(countFile, sampleData, design = ~sample + exon +
  condition:exon, flattenedfile = NULL)
}
\description{
Load Fcount output from : DEXSeq_after_Fcount.R into DEXSeq
 Copyright 2015 Vivek Bhardwaj (bhardwaj@ie-freiburg.mpg.de). Licence: GPLv3.
 Read Fcount output and convert to dxd
}
