% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logTime.R
\name{logTime}
\alias{logTime}
\title{Load data from OfficeTime export into my time recording DB
Takes a file name and squirts the info into the time recording database. It will check if
data already exists for the dates and stop rather than write duplicate entries.}
\usage{
logTime(
  fileName = NULL,
  projDir = "/Users/pschofield/Projects/time_sheet/",
  dbName = "timeSheet.db"
)
}
\arguments{
\item{fileName}{name of officetime output file}

\item{projDir}{default directory of the project}

\item{dbName}{default name of the timesheet sqlite database}
}
\description{
Load data from OfficeTime export into my time recording DB
Takes a file name and squirts the info into the time recording database. It will check if
data already exists for the dates and stop rather than write duplicate entries.
}
