% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{macOpen}
\alias{macOpen}
\title{simple wrapper to open a file at the commandline on a mac}
\usage{
macOpen(fileName)
}
\arguments{
\item{fileName}{name of the file}
}
\description{
very simple command possibly expand to take application as an argument
}
