% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBAMS.R
\name{mergeBAMS}
\alias{mergeBAMS}
\title{merge some bamfiles}
\usage{
mergeBAMS(sampleName, bamDir, outDir, projName, noSub = F, ncores = 16)
}
\arguments{
\item{sampleName}{stub of the sample name}

\item{bamDir}{the directory where the files are}

\item{outDir}{where to put the merged files}
}
\description{
merge some bamfiles
}

