% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myip.R
\name{myip}
\alias{myip}
\title{myip}
\usage{
myip(pattern = "Bcast", command = "ifconfig", item = "addr",
  linesep = "  ", namesep = ":")
}
\arguments{
\item{pattern}{pattern to search for in command output}

\item{command}{system command to produce ip configuration output}

\item{item}{name of first item to trim line to}

\item{linesep}{pattern between items on address line}

\item{namesep}{pattern to split items into name and address}
}
\description{
function to return the localhost's ip addresses
}
