% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newFile.R
\name{newFile}
\alias{newFile}
\title{create a new notebook file}
\usage{
newFile(
  project = NULL,
  fname = NULL,
  root = "/Users/pschofield/git_tree/",
  template = "/Users/pschofield/git_hub/template.Rmd"
)
}
\arguments{
\item{root}{default project root directory}

\item{name}{project name}
}
\description{
create a new notebook file
}
