% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newNote.R
\name{oldNote}
\alias{oldNote}
\title{create a new notebook file}
\usage{
oldNote(name = NULL, root = "/Users/pschofield/Code/")
}
\arguments{
\item{name}{project name}

\item{root}{default project root directory}
}
\description{
create a new notebook file
}
