% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newProject.R
\name{oldProject}
\alias{oldProject}
\title{create a new project}
\usage{
oldProject(
  name = NULL,
  root = "/Users/pschofield/Code/",
  noidx = F,
  noDir = F,
  noWWW = F
)
}
\arguments{
\item{name}{name of project}

\item{root}{root directory for project directory}

\item{noidx}{don't create a default index.Rmd file}

\item{noDir}{don't create an output directory}

\item{noWWW}{don't create a soft link under public_html on server}
}
\description{
create a new project
}
