% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendb.R
\name{openDB}
\alias{openDB}
\title{open a database using the credentials file}
\usage{
openDB(db, credir = getwd(), credfile = file.path(credir,
  ".psqlcred.yml"))
}
\description{
open a database using the credentials file
}
