% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plib.R
\docType{package}
\name{plib}
\alias{plib}
\alias{plib-package}
\title{plib}
\description{
A library of convenience functions to make life easier while produce workday notebooks to record what
I have been doing inparticular submitting and running processing jobs on the cluster. This library
used to exist as two separate libraries, my original pietalib that I started at Dundee and a CRUKlib
that I started at CRUK Manchester. However it makes more sense to combine these now.
}

