% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plibr.R
\docType{_PACKAGE}
\name{plibr}
\alias{plibr}
\title{plibr}
\description{
Latest iteration of my convenience function code for managing Rmarkdown wrangling for creating
electronic notebooks
}
\details{
A library of convenience functions to make life easier while produce workday notebooks to record what
I have been doing in particular submitting and running processing jobs on the cluster. This library
used to exist as two separate libraries, my original pietalib that I started at Dundee and a CRUKlib
that I started at CRUK Manchester. However it makes more sense to combine these now.

Most of this is redundant code now but I am loathed to lose it certainly all the batch-job submission
management for bioinformatics jobs is not distant history
}
