% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\title{Plot and optionally save a figure depending on knitting context}
\usage{
plot_save(
  plot,
  filename = NULL,
  outdir = "figures",
  width = 12,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{plot}{A ggplot object to plot.}

\item{filename}{Optional base filename (without extension). Defaults to the chunk label if available.}

\item{outdir}{Directory to save plots into (created if missing). Default is "figures".}

\item{width}{Plot width in inches. Default is 12.}

\item{height}{Plot height in inches. Default is 8.}

\item{dpi}{Resolution for saved images. Default is 300.}
}
\value{
Invisibly returns the input plot.
}
\description{
This helper function plots a ggplot object and, if called during
R Markdown rendering, saves it automatically to a specified folder.
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
plot_save(p)

}
