% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNV.R
\name{plotCNV}
\alias{plotCNV}
\title{Plot CNV profiles}
\usage{
plotCNV(copyData, segData, paramList, reduceCol = T)
}
\arguments{
\item{copyData}{list of RangedData Objects with copy number}

\item{segData}{list of corresponding segmentation data}

\item{param}{list of initial parameter}
}
\description{
There are 6 states, which in a diploid sample corresponds to the following
chromosomal copies and biological state:
 1  <= 0 copies, homozogous deletion
 2  == 1 copy, heterozogous deletion
 3  == 2 copies, neutral
 4  == 3 copies, gain
 5  == 4 copies, amplification
 6  >= 5 copies, high level amplification
Output:
states - The state assigned to each copy number value
segs  - Non-overlapping segments and medians of each segment
mus  - Optimal median of of copy numbers in state
lambda - Optimal precision of copy numbers in state
pi   - Optimal state distribution
loglik - The likelihood values of each EM algorithm iteration
rho  - Posterior marginals (responsibilities) for each position and state
}
