% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSubTree.R
\name{plotSubTree}
\alias{plotSubTree}
\title{plot a dendrogram borrowed from ape zoom with ammendments}
\usage{
plotSubTree(phy, focus, subtree = FALSE, col = rainbow, ...)
}
\description{
plot a dendrogram borrowed from ape zoom with ammendments
}
