% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\title{Publish all relevant files from the local directory to my public_html tree}
\usage{
publish(rname = "/Users/pschofield/Code/", pname = NULL,
  fname = NULL, note = T, outStub = NULL, delete = F, open = T)
}
\arguments{
\item{note}{publish to the notebook directory}

\item{outStub}{name of subdirectory to publish to}

\item{delete}{delete the local files after copying}

\item{open}{open the file on publishing}

\item{filePath}{full name of the markdown file from which the html was generated}
}
\description{
Publish all relevant files from the local directory to my public_html tree
}
