% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstat.R
\name{qstat}
\alias{qstat}
\title{qstat}
\usage{
qstat(user = "pschofield")
}
\arguments{
\item{user}{the username for the cluster job listing}
}
\description{
run qstat on the cluster remotely and return the results to an R session this is
currently just a simple function which defaults to my user name
}
