% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rcur}
\alias{rcur}
\title{render the current document}
\usage{
rcur(fileName = .curFile, projDir = .curProj, inDir = NULL,
  outDir = NULL, open = "html", godPath = "public_html/work/Projects",
  rootDir = file.path(Sys.getenv("HOME"), "Projects"),
  codeDir = file.path(Sys.getenv("HOME"), "Projects"),
  sysId = Sys.info()["sysname"], setGH = F, htmlApp = "google-chrome",
  pdfApp = "evince", wordApp = "loffice", sourcecopy = F, toPDF = F,
  toDOCX = F, toHTML = T, toGOD = T, GOD = T, setWH = F)
}
\arguments{
\item{fileName}{file name}

\item{projDir}{project name}

\item{inDir}{override codeDir construction with exact path}

\item{outDir}{copy to specific output subdirectory}

\item{open}{view the output file}

\item{rootDir}{root of output tree}

\item{codeDir}{code tree path}

\item{setWH}{sets this file as the home file in the work directory}

\item{sourceCopy}{copy source to the output location too}

\item{toData}{make a copy of the output in another public location}

\item{dataDir}{secondary location for copy}

\item{locOnly}{write output to local hardrive rather than network share}
}
\description{
renders the named document and puts the output html file in the correct location in my filespace
This currently assumes that I have mounted my troodon home directory to the local host via sshfs
which is available from OSXFUSE. Default paths are configured for my code locations.
}
\details{
Been making changes to this to use bookdown document styles
}
