% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDREME.R
\name{readDREME}
\alias{readDREME}
\title{read in dreme results file}
\usage{
readDREME(filename)
}
\arguments{
\item{filename}{name of xml file}
}
\value{
returns a list with an array of statistic on each motif and an list of all the pwms
}
\description{
read in dreme results file
}

