% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readENCODE.R
\name{readENCODE}
\alias{readENCODE}
\title{readENCODE read broadPeak and narrowPeak formats and return a GRanges object}
\usage{
readENCODE(filename)
}
\arguments{
\item{filename}{filename to import}
}
\description{
readENCODE read broadPeak and narrowPeak formats and return a GRanges object
}

