% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFC.R
\name{readFC}
\alias{readFC}
\title{readFC read a featureCount File}
\usage{
readFC(filename, exons = F, name = gsub("[.]bam$", "",
  basename(filename)))
}
\arguments{
\item{filename}{name of feature counts output file}

\item{exons}{boolean if features are exons and there will be duplicate gene ids}
}
\description{
readFC read a featureCount File
}
