% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMEME.R
\name{readMEME}
\alias{readMEME}
\title{read in meme xml results file into an R list structure}
\usage{
readMEME(filename)
}
\arguments{
\item{filename}{name of xml file}
}
\value{
returns a list with an array of statistic on each motif and an list of all the pwms
}
\description{
read in meme xml results file into an R list structure
}
