% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPrism.R
\name{readPrism}
\alias{readPrism}
\title{Read a Prism file.}
\usage{
readPrism(filename)
}
\arguments{
\item{filename}{the name of the prism file}
}
\description{
This function is in no way a generic Prism reader it is specifically designed
to read qRT_PCR results files from A Seifert
}

