% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceevery.R
\name{replace_every}
\alias{replace_every}
\title{replace every nth character in a string with a character}
\usage{
replace_every(stringin, enth, ch)
}
\arguments{
\item{stringin}{the original string}

\item{enth}{the length of the repeat}

\item{ch}{the replacement character}
}
\description{
replace every nth character in a string with a character
}
