% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsh.R
\name{rsh}
\alias{rsh}
\title{execute a command remotely via ssh}
\usage{
rsh(cmd, hostname = "feenix", username = "pschofield")
}
\arguments{
\item{cmd}{the command to execute}

\item{hostname}{the remote machine}

\item{username}{currently works with ssh keys rather than password}
}
\description{
execute a command remotely via ssh
}
