% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFC.R
\name{runFC}
\alias{runFC}
\title{Run featureCounts on a set of bam files}
\usage{
runFC(projName, bamDir, fileList, outFile, ncores = 16,
  subreadMod = "apps/subread/1.5.0-p3/gcc-4.4.7",
  refDir = "/scratch/pschofield/ref/star/GRCh37_pa",
  annoFile = file.path(refDir, "Homo_sapiens.GRCh37.85.gtf"),
  genomeFile = file.path(refDir,
  "Homo_sapiens.GRCh37.dna_sm.primary_assembly.fa"),
  switches = " -M -O -C -P -p -t exon -g transcript_id -f -d 50")
}
\arguments{
\item{projName}{project name}

\item{bamDir}{location of bamfiles (actually directory where counting takes place)}

\item{fileList}{list of files if not full paths will have to be in bamDir but can be fully qualified
paths}

\item{outFile}{name of output file}

\item{ncores}{number of cores to use}

\item{annoFile}{the gtf annotation file to count against must of course match the alignment reference}

\item{switches}{extra switches for featureCounts for example paired end or exon counting}
}
\description{
Run featureCounts on a set of bam files
}

