% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHMMcopy.R
\name{runHMMcopy}
\alias{runHMMcopy}
\title{Script for performing copy number analysis using HMMcopy}
\usage{
runHMMcopy(projName, pairings = NULL, wigDir,
  remRoot = "/scratch/pschofield/Projects", mapFile, gcFile, paramList,
  chrToPlot = c(1:22, "X", "Y"))
}
\description{
adapted from a script created: 11 Aug 2016, LHS
}
