% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runJAFFA.R
\name{runJAFFA}
\alias{runJAFFA}
\title{run JAFFA fusion gene detection}
\usage{
runJAFFA(sampleId, dataDir, filePattern, projName, outDir,
  refGenome = "hg19", refAnno = "genCode19", readLayout = "single",
  javaMod = "compilers/java/1.8.0_40/noarch",
  velvetMod = "apps/velvet/1.2.10/gcc-4.4.7",
  oasesMod = "apps/oases/0.2.08/gcc-4.4.7",
  blatMod = "apps/blat/35/gcc-4.4.7",
  samtoolsMod = "apps/samtools/0.1.19/gcc-4.4.7",
  bowtie2Mod = "apps/bowtie2/2.2.1/gcc/4.4.7",
  rMod = "apps/R/3.3.0/gcc-4.8.5",
  jaffaPath = "/scratch/pschofield/Projects/Sandbox/JAFFA-version-1.08",
  ncores = 2, mem = "32Gb", noSub = F)
}
\arguments{
\item{sampleId}{id for naming the job script}

\item{dataDir}{location of the fastq files}

\item{filePattern}{file pattern to include all the relevant fastq files}

\item{projName}{name of the project}

\item{outDir}{full path to file output}
}
\description{
run JAFFA fusion gene detection
}
