% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMACS2.R
\name{runMACS2}
\alias{runMACS2}
\title{submit a MACS2 job on the the cluster}
\usage{
runMACS2(trBam, ctBam, outDir, name, broad = T, pe = F, projName,
  genome = "hs", bcutoff = 0.01, pyMod = "apps/python",
  pyLib = "libs/python2sitepkgs", noMod = F,
  remRoot = file.path("/scratch/pschofield/Projects", projName))
}
\arguments{
\item{trBam}{treatment bam file}

\item{ctBam}{control bam file}

\item{outDir}{output directory}

\item{name}{output name}

\item{broad}{call broad peaks}
}
\description{
submit a MACS2 job on the the cluster
}
