% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQSTrim.R
\name{runQSTrim}
\alias{runQSTrim}
\title{Submit a list of files associated with a project for  QC checking with fastQC file}
\usage{
runQSTrim(projName, outRoot = file.path("/scratch/pschofield/Projects",
  projName), fileList = NULL, outName = "Data/trimmedfqs", noSub = F)
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
Submit a list of files associated with a project for  QC checking with fastQC file
}
