% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQualimap.R
\name{runQualimap}
\alias{runQualimap}
\title{QC check using Qualimap}
\usage{
runQualimap(sampleID, projName, bamDir, bamExt = "_dedup_realn_recal.bam",
  outDir = file.path("/scratch/pschofield/Projects", projName,
  "Analysis/qualimap"), noSub = F, ncores = 1, scpIt = T, mem = "32Gb",
  qualimapMod = "apps/qualimap/2.2/linux-x86_64")
}
\arguments{
\item{fileList}{the bam files}
}
\description{
QC check using Qualimap
}
