% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRScript.R
\name{runRScript}
\alias{runRScript}
\title{generate and submit a batch job that runs an Rjob}
\usage{
runRScript(rscript, args = NULL, jname, jproj, jdesc,
  remRoot = "/scratch/pschofield/Projects",
  locRoot = "/Users/pschofield/Projects", ncores = 1, nproc = 8,
  mem = "16Gb", wtime = "30:00:00", overwrite = T, eviron = F,
  rMod = "apps/R/3.3.0/gcc-4.8.5")
}
\arguments{
\item{args}{any additional arguments}

\item{jname}{job name}

\item{jproj}{project name}

\item{jdesc}{job description}

\item{rMod}{which version of R to run}

\item{sname}{name of the script file}

\item{logdir}{log directory}

\item{locroot}{local root}

\item{remroot}{remote root}
}
\description{
generate and submit a batch job that runs an Rjob
}
