% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSICER.R
\name{runSICER}
\alias{runSICER}
\title{run SICER on a set of file pairs}
\usage{
runSICER(projName, filePairs, bedDir, noControl = F, pyMod = "apps/python",
  pyLib = "libs/python2sitepkgs",
  sicerPath = "/scratch/pschofield/local/SICER",
  outRoot = "/scratch/pschofield/Projects", outPath = "", winsize = 500,
  gapsize = 4000, genome = "GRCh37", fdr = "1E-3", insize = 300,
  gencov = 0.8)
}
\arguments{
\item{projName}{name of the project}

\item{filePairs}{list of file pairs}
}
\description{
run SICER on a set of file pairs
}
