% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runStarAlign.R
\name{runStarAlign}
\alias{runStarAlign}
\title{Submit a list of files associated with a project for bwa alignment and deduping}
\usage{
runStarAlign(projName, sampleID, fileTable, runName,
  outRoot = file.path("/scratch/pschofield/Projects", projName),
  outName = "Data/alignment/star", noSub = F, scpIt = T,
  starMod = "apps/star/2.5.1b/gcc-5.1.0",
  samtoolsMod = "apps/samtools/0.1.19/gcc/4.4.7", refSequence = NULL,
  refGenome = NULL, ncores = 16, fqExt = ".fq", pe = T, jm = F)
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
Submit a list of files associated with a project for bwa alignment and deduping
}

