% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortBAM.R
\name{sortBAM}
\alias{sortBAM}
\title{sort a bamfile}
\usage{
sortBAM(inBAM, outSuf = NULL, projName, noSub = F, ncores = 16)
}
\arguments{
\item{inBAM}{name of input file}

\item{outBAM}{name of sorted file if NULL inBAM is overwritten}
}
\description{
sort a bamfile
}
