% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitBam.R
\name{splitBAMS}
\alias{splitBAMS}
\title{a routine to split and filter by insert size strand specific bam files}
\usage{
splitBAMS(bamFiles, len = 150, projName, outDir = NULL, stranded = 1)
}
\arguments{
\item{bamFiles}{and list of fully qualified bamfile names}

\item{len}{length of insert to split by}
}
\description{
a routine to split and filter by insert size strand specific bam files
}
