% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitStrandedBam.R
\name{splitStrandedBAM}
\alias{splitStrandedBAM}
\title{a routine to split and filter by insert size strand specific bam files}
\usage{
splitStrandedBAM(bamFiles)
}
\arguments{
\item{bamFiles}{and list of fully qualified bamfile names}

\item{len}{length of insert to split by}
}
\description{
a routine to split and filter by insert size strand specific bam files
}
