% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_tidy.R
\name{sql_tidy}
\alias{sql_tidy}
\title{tidy up the sql}
\usage{
sql_tidy(
  sql,
  prefix = "tmp",
  sql_keywords = c("SELECT", "FROM", "WHERE", "AND", "OR", "GROUP", "ORDER", "HAVING",
    "INSERT", "VALUES", "UPDATE", "SET", "DELETE", "CREATE", "DROP", "JOIN", "WITH",
    "INNER", "LEFT", "RIGHT", "OUTER", "FULL", "CROSS", "/\\\\*", "\\\\*/", "--")
)
}
\description{
tidy up the sql
}
