% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandedBamImport.R
\name{strandedBamImport}
\alias{strandedBamImport}
\title{Stranded bam import}
\usage{
strandedBamImport(file, selection)
}
\description{
A function to extract reads from a stranded RNA-seq BAM file.  Intended to be used as a custom
importFunction in the DataTrack constructor of Bioconductor's Gviz package. With properly
stranded data this allows the plotting of coverage tracks from each strand of the genome.
6 Nov 2014 - bensidders at gmail dot com
UPDATE: 10 Sep 2015: The funciton can now plot data from libraries created in either
orientation. Due to the manner in which Gviz calls §this function, you need to specify the
library orientation with a global variable "libType" to be one of "fr-firststrand" or
"fr-secondstrand".
}
