% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringtoasc.R
\name{string_to_seed}
\alias{string_to_seed}
\title{convert a string to a number that can be used as a seed for random}
\usage{
string_to_seed(stringin)
}
\description{
convert a string to a number that can be used as a seed for random
}
