% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subFastQC.R
\name{subFastQC}
\alias{subFastQC}
\title{Submit a cluster job to fastQC a directory}
\usage{
subFastQC(fastqDir, fastqPattern = "*.gz", outDir = "pre",
  projName = "Sandbox", noSub = F,
  projDir = file.path("/scratch/pschofield/Projects", projName),
  fastQCDir = file.path("/scratch/pschofield/Projects", projName,
  "fastqc"), jobName = paste0("fastQC_", projName),
  fastqcMod = "apps/fastqc/0.11.3/linux-x86_64", nproc = 16,
  mem = "48gb", makeCVS = T)
}
\arguments{
\item{fastqDir}{input directory}

\item{fastqPattern}{pattern for fastq file names unix wildcards not re}

\item{outDir}{subdirectory of project fastqc directory}

\item{projName}{project name will build output path}

\item{noSub}{debug flag}

\item{fastQCDir}{output root directory}

\item{jobName}{name for job script}

\item{fastqcMod}{fastQC module to load}

\item{makeCVS}{make cvs file with listing details}
}
\description{
Submit a cluster job to fastQC a directory
}
