% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subScript.R
\name{subScript}
\alias{subScript}
\title{subScript will submit a script to the cluster}
\usage{
subScript(scriptstub = "pieta_qsub", script = c("#!/bin/bash", "hostname"),
  scriptext = ".sh", tmpdir = "/homes/pschofield/tmp/", logdir = "",
  cores = 8, nosub = FALSE, email = "as", queue = NULL,
  emailAddress = "p.schofield@dundee.ac.uk", args = NULL, ramSize = "4G",
  nNodes = 1)
}
\arguments{
\item{scriptstub}{the stub for the temporary file name}

\item{script}{the content of the script as a vector of strings}

\item{scriptext}{extention for the temporary file}

\item{tmpdir}{location for the temporary file}

\item{logdir}{location for the batch job logs}

\item{cores}{number of cores}

\item{nosub}{boolean to switch off actual submission but to create the tempfiles}

\item{email}{job status email settings}

\item{queue}{queue name}

\item{emailAddress}{where to send job status emails}

\item{args}{any arguments to add to the job submission string}
}
\description{
calls subJob to submit a script to the cluster as a temporary file it relies on the temporary
directory where the temporary file will be written to being mounted to the local machine.
I could get round this by writing it locally and then copying it with scp but at the moment
this is not worth the effort.
}
