% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumCNValignStats.R
\name{sumCNValignStats}
\alias{sumCNValignStats}
\title{Summarise the mapping statistics information for a specific dataset}
\usage{
sumCNValignStats(projName,
  remRoot = file.path("/scratch/pschofield/Projects", projName),
  mapStatDir = "Analysis/mapping_stats", miSeq = TRUE,
  locDir = file.path("/Users/pschofield/Projects", projName, "Data"),
  asFilePat = "_align.stats", fsFilePat = "_insertSize.txt")
}
\arguments{
\item{projName}{project name}

\item{bamDir}{bam alignment stats directory}

\item{outDir}{output directory}
}
\description{
Summarise the mapping statistics information for a specific dataset
}
