% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumForPlot.R
\name{sumForPlot}
\alias{sumForPlot}
\title{sumForPlot}
\usage{
sumForPlot(dfin, cols, val)
}
\arguments{
\item{dfin}{data frame in}

\item{cols}{over}

\item{val}{field}
}
\description{
generate a summary for creating pretty boxplots for data with missing values
}
