% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseQualiMap.R
\name{summariseQualimap}
\alias{summariseQualimap}
\title{gather alignment statistic on a directory of star alignments}
\usage{
summariseQualimap(fileList, projName, locDir = "/Users/pschofield/.tmp")
}
\arguments{
\item{fileList}{list of log files}

\item{projName}{project names}

\item{locDir}{location for (temporary) local storage of log files}
}
\description{
gather alignment statistic on a directory of star alignments
}
