% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSTAR.R
\name{summariseSTAR}
\alias{summariseSTAR}
\title{gather alignment statistic on a directory of star alignments}
\usage{
summariseSTAR(projName, fileList)
}
\arguments{
\item{projName}{project names}

\item{fileList}{list of log files}

\item{locDir}{location for (temporary) local storage of log files}
}
\description{
gather alignment statistic on a directory of star alignments
}
