% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSubread.R
\name{summariseSubread}
\alias{summariseSubread}
\title{gather alignment statistic on a directory of star alignments}
\usage{
summariseSubread(projName, fileList, locDir = "/Users/pschofield/.tmp")
}
\arguments{
\item{projName}{project names}

\item{fileList}{list of log files}

\item{locDir}{location for (temporary) local storage of log files}
}
\description{
gather alignment statistic on a directory of star alignments
}

