% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survAnal.R
\name{survAnal}
\alias{survAnal}
\title{survAnal: runs cox proportional hazard models against patients with top and bottom n percent
expression values}
\usage{
survAnal(survData = NULL, exprData = NULL, geneData = NULL,
  survTime = "time", exprRange = 0.33, survStatus = "status",
  mcores = 4, summarise = TRUE)
}
\arguments{
\item{survData}{survival data must contain a time and status}

\item{exprData}{normalised gene expression data}

\item{geneData}{names of the genes to be tested (must be in the rownames of exprData)}

\item{survTime}{name of survival time column}

\item{exprRange}{top and bottom quantiles of expression values to select samples for testing}

\item{survStatus}{name of suvival censoring column}

\item{mcores}{number of cores for bioparallel}
}
\value{
the function creates several csv output files with the Cox PH results, the Limma DGE
results, a combined set of both results and a set of those genes for which the DGE result is
significant at 0.05 FDR
}
\description{
derived from the runCPH function will process three separate file one with sample ids one with
survival data and one with expression values. In test mode if a list of sample ids are supplied
but not the other two files it assumes the ids are from the TCGA SKCM data set and uses a
cached version of the normalised expression values from the TCGA SKCM data set and a currated
set of survival data.
}
