% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survClustAnal.R
\name{survClustAnal}
\alias{survClustAnal}
\title{survival analysis of expression profile clustered data}
\usage{
survClustAnal(ed, gd, pd, trim = 0.25, scale = "col", th = 50,
  dRow = "canberra", cRow = "average", scaleIt = T, cCol = "complete",
  dCol = "correlation", st = "time", ss = "status", rowA = NULL,
  rcuts = (nrow(pd)/10), ccuts = (length(gd)/10))
}
\description{
survival analysis of expression profile clustered data
}
