% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFit.R
\name{survFit}
\alias{survFit}
\title{survAnal: runs cox proportional hazard models against patients with top and bottom n percent
expression values}
\usage{
survFit(survData = NULL, exprData = NULL, geneData = NULL,
  survTime = "time", survStatus = "status", mcores = 4,
  verbose = F)
}
\arguments{
\item{survData}{survival data must contain a time and status}

\item{exprData}{normalised gene expression data}

\item{geneData}{names of the genes to be tested (must be in the rownames of exprData)}

\item{survTime}{name of survival time column}

\item{survStatus}{name of suvival censoring column}

\item{mcores}{number of cores for bioparallel}

\item{exprRange}{top and bottom quantiles of expression values to select samples for testing}
}
\value{
the function creates several csv output files with the Cox PH results, the Limma DGE
results, a combined set of both results and a set of those genes for which the DGE result is
significant at 0.05 FDR
}
\description{
The runCPH function will process three separate file one with sample ids one with survival data
and one with expression values. In test mode if a list of sample ids are supplied but not the
other two files it assumes the ids are from the TCGA SKCM data set and uses a cached version of
the normalised expression values from the TCGA SKCM data set and a currated set of survival
data.
}
\examples{
analysisName="test3"
exprSep=","
res <- runCPH(sampleDataFile=sampleDataFile,expressionDataFile=expressionDataFile,
              survivalDataFile=survivalDataFile,
              analysisName=analysisName,exprSep=exprSep)

}
