% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixFile.R
\name{tabixFile}
\alias{tabixFile}
\title{run tabix to produce compressed indexed file}
\usage{
tabixFile(infile, outfile = NULL, outpath = NULL, form = NULL, sortstr = NULL)
}
\arguments{
\item{infile}{input file}

\item{outfile}{output file}

\item{outpath}{output directory path}
}
\description{
A simple function to tabix bed and gtf files it might work for other files but I have not tried any so
is likely that it will fail.
}
