% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark_image.R
\name{watermark_image}
\alias{watermark_image}
\title{Water mark and scale image}
\usage{
watermark_image(
  img_path,
  text = "© Pietà Georgie Schofield ",
  opacity = 0.3,
  width = 800,
  ret = NULL,
  saveimg = NULL
)
}
\description{
Water mark and scale image
}
