% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{build_project_index}
\alias{build_project_index}
\title{build project index}
\usage{
build_project_index(
  project,
  git_directory = file.path(Sys.getenv("HOME"), "repositories"),
  html_directory = file.path("/srv", "http"),
  repo = "uol",
  coderoot = NULL,
  htmlroot = "uol",
  pub = F
)
}
\description{
build project index
}
