% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAll.R
\name{buildAll}
\alias{buildAll}
\title{Build an Rmarkdown file or directory of files into html}
\usage{
buildAll(
  project,
  file = "*",
  outstub = "",
  outdir = "/homes/pschofield/",
  indir = "/Users/pschofield/git_tree/",
  explicit = FALSE,
  outType = "html_document",
  interact = FALSE
)
}
\arguments{
\item{project}{the name of the project}

\item{file}{the name of the file (this is the file stub without path or extention)}

\item{outdir}{the location where the output file will be written}

\item{indir}{the path to the input file}

\item{explicit}{boolean specifies that the paths will be use rather than constructed}

\item{outType}{character string type of output 'all' will make all}
}
\description{
This function is to make it easier to build pages from Rmd files in my personal webspace
It is a bit of a work in progress I am trying to keep it logical and not have too many
exceptions that need special cases. At the moment I have not sorted out index pages for
the private and public directories
}
