% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data.R
\name{display_data}
\alias{display_data}
\title{display datatable with different table types}
\usage{
display_data(
  dataset,
  number = NULL,
  table_type = "DT",
  limited = FALSE,
  row_names = FALSE,
  buttons = FALSE,
  plen = NULL,
  caption = NULL,
  dbug = FALSE,
  fixh = TRUE,
  fixc = list(leftColumns = 1),
  sigf = 3
)
}
\arguments{
\item{dataset}{the data to display}

\item{number}{the maximum number of rows to display}

\item{table_type}{the type of table to use ("DT", "kable", "flextable")}

\item{limited}{constrain to container}

\item{row_names}{display row names}

\item{buttons}{include export buttons}

\item{plen}{page length for DT}

\item{caption}{table caption}

\item{dbug}{debugging flag}

\item{fixh}{fix header}

\item{fixc}{fix column options}

\item{sigf}{significant figures for numeric formatting}
}
\description{
display datatable with different table types
}
