% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_data.R
\name{display_data}
\alias{display_data}
\title{display datatable}
\usage{
display_data(
  dataset,
  number = NULL,
  disp = T,
  limited = F,
  buttons = F,
  plen = NULL,
  caption = NULL,
  fixh = T,
  fixc = list(leftColumns = 1),
  sigf = 3
)
}
\arguments{
\item{dataset}{the data to display}

\item{number}{the maxium number of rows to display}

\item{disp}{the flag which decides the whole shebang}

\item{limited}{contrain to container}
}
\description{
displays a datatable (DT) either as raw tibble if not compiling to avoid trying to fire up unnecessary
browers or as a DT if rendering to html
}
