% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCluster.R
\name{evalCluster}
\alias{evalCluster}
\title{evalCluster}
\usage{
evalCluster(expData, survData, hm = F, gcl = F, gcts = 10,
  cmeth = "ward.D2")
}
\description{
function from etv1 projects not totally sure when I wrote this I have no memory of it
}
