% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{get files from the cluster using scp to avoid using fuse for the moment}
\usage{
getFiles(
  fn = NULL,
  filenames = NULL,
  projName,
  locDir = NULL,
  recursive = F,
  host = "god",
  user = "pieta",
  force = F,
  simple = T
)
}
\arguments{
\item{locDir}{local directory path}

\item{host}{hostname}

\item{user}{username}

\item{filename}{base files}

\item{remDir}{remote directory path}

\item{remRoot}{remote root}

\item{locRoot}{local root}
}
\description{
get files from the cluster using scp to avoid using fuse for the moment
}
