% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles.R
\name{getFiles}
\alias{getFiles}
\title{get files from the cluster using scp to avoid using fuse for the moment}
\usage{
getFiles(filenames, remDir, locDir = remDir, remRoot = "/scratch",
  locRoot = "/Users", host = "troodon", user = "pschofield", force = F)
}
\arguments{
\item{remDir}{remote directory path}

\item{locDir}{local directory path}

\item{remRoot}{remote root}

\item{locRoot}{local root}

\item{host}{hostname}

\item{user}{username}

\item{filename}{base files}
}
\description{
get files from the cluster using scp to avoid using fuse for the moment
}

