% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmAnalysis.R
\name{hmAnalysis}
\alias{hmAnalysis}
\title{use heatmap and tree cutting}
\usage{
hmAnalysis(exprMat, rowV = c("correlation", "mcquitty"),
  colV = c("canberra", "average"), ncuts = 6, filename = NA,
  mainTitle = "Heatmap")
}
\description{
use heatmap and tree cutting
}
