% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmAnalysis.R
\name{hmAnalysis}
\alias{hmAnalysis}
\title{use heatmap and tree cutting}
\usage{
hmAnalysis(exprMat, rowV = NULL, colV = NULL, height = NA, width = NA,
  ch = NA, cw = NA, leg = T, al = T, ncuts = NULL, filename = NA,
  mainTitle = "Heatmap", cr = 1.5, cc = 0.5, scale = scale)
}
\description{
use heatmap and tree cutting
}
