% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlm_without_rmd.R
\name{html_without_rmd}
\alias{html_without_rmd}
\title{List HTML files without corresponding Rmd source files}
\usage{
html_without_rmd(
  projName,
  rmd_dir = .codeDir,
  html_dir = file.path("/srv", "http", "uol", projName),
  dbug = FALSE
)
}
\arguments{
\item{rmd_dir}{Path to the directory containing \code{.Rmd} files.}

\item{html_dir}{Path to the directory containing \code{.html} files.}

\item{as_url}{Logical. If TRUE, \code{linkaddress} is returned as an HTTP-style
path using \code{base_url}. If FALSE (default), file paths are used.}

\item{base_url}{Optional base URL to prepend when \code{as_url = TRUE}.}
}
\value{
A tibble with two columns:
\describe{
\item{name}{Base name of the HTML file (without extension).}
\item{linkaddress}{Either a file path or an HTTP link to the HTML file.}
}
}
\description{
Scans two directories: one containing rendered \code{.html} files, and one
containing source \code{.Rmd} files. Returns a tibble of \code{.html} files that do not
have a matching \code{.Rmd} (same base name) in the code directory.
}
\examples{
list_html_without_rmd(
  html_dir = "/srv/http/uol/project",
  rmd_dir = "/home/me/repositories/uol/project"
)

}
