% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmonth.R
\name{newNote}
\alias{newNote}
\title{make a new notebook for a project}
\usage{
newNote(
  fileName,
  projName,
  rootDir = file.path(Sys.getenv("HOME"), "Repos"),
  gitRepo = "uol",
  ext = "Rmd",
  db = F,
  template = "notepage",
  package = "plibr",
  chkidx = T,
  ed = F
)
}
\description{
make a new notebook for a project
}
