% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmonth.R
\name{newnote}
\alias{newnote}
\title{make a new notebook for a project}
\usage{
newnote(
  fileName,
  projName,
  rootDir = file.path(Sys.getenv("HOME"), "GitLab"),
  gitRepo = "liverpool",
  ext = "Rmd",
  db = F,
  template = "notepage",
  package = "plibr",
  chkidx = T,
  ed = F
)
}
\description{
make a new notebook for a project
}
