% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opendb.R
\name{openDB}
\alias{openDB}
\title{open a database using the credentials file}
\usage{
openDB(db, credir = NULL, credfile = "psqlcred.yml")
}
\arguments{
\item{db}{the name of the database}

\item{credir}{the location of the credentials file}

\item{credfile}{the name of the credentials file if not standard}
}
\description{
open a database using the credentials file
}
