% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myKMplot.R
\name{plotKM}
\alias{plotKM}
\title{plot Kaplan Myer curves}
\usage{
plotKM(formstr, dat, alp = I(1/10), sfact = "surv", inclusive = F,
  y1 = -0.5, y2 = -1, titlestr = gsub(".*[~]", "", formstr))
}
\arguments{
\item{alp}{alpha for transparancy of errors}

\item{sfact}{survival value to plot}

\item{fit}{result of survfit}

\item{cox}{result of coxph}
}
\description{
plot Kaplan Myer curves
}
