% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\name{rb}
\alias{rb}
\title{rebuild a package}
\usage{
rb(
  wd = NULL,
  projName = "plibr",
  pkgName = "package",
  install = F,
  baseDir = file.path(Sys.getenv("HOME"), "Repos")
)
}
\arguments{
\item{pkgname}{the package name}
}
\description{
rebuild a package
}
