% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\name{rb}
\alias{rb}
\title{rebuild a package}
\usage{
rb(wd = NULL, projName = "plib_ps", pkgName = "package",
  baseDir = file.path(Sys.getenv("HOME"), "Projects"))
}
\arguments{
\item{pkgname}{}

\item{rstudio}{}
}
\description{
rebuild a package
}
