% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rc}
\alias{rc}
\title{render the current document}
\usage{
rc(
  fileName = .curFile,
  projName = .projName,
  gitRepo = .gitRepo,
  sysRoot = Sys.getenv("HOME"),
  user = Sys.getenv("USER"),
  codeDir = file.path(sysRoot, "GitLab", gitRepo),
  hostname = "dh174037.liv.ac.uk",
  outRoot = file.path(sysRoot, ".tmp"),
  htmlRoot = file.path("/", "var", "www", "html"),
  shinyRoot = file.path("/", "u1", "shiny-server", "samples", "apps"),
  onedrive = file.path(sysRoot, "OneDrive", "me", "html"),
  docRoot = file.path(sysRoot, "Projects", projName),
  setHome = F,
  toPDF = F,
  toDOCX = F,
  toHTML = T,
  setRepo = T,
  setProj = T,
  toShiny = F,
  htmlUP = T,
  rmdUP = F,
  pdfUP = F,
  docUP = F,
  ext = "Rmd",
  dbg = F
)
}
\arguments{
\item{fileName}{file name}

\item{sysRoot}{home directory}

\item{codeDir}{code tree path}

\item{outRoot}{local temporary file creation location}

\item{projDir}{project name}

\item{userid}{User id}
}
\description{
This function has evolved way beyond the parameter list. It has also had to adapt from the original
dundee version to a version that worked with GoDaddy and then worked at Manchester CRUK and then cope
with the changes at CRUK and the whole reconfiguration due to the fire and then the  move to Liverpool
and the introduction of the virtual machine at Liverpook running a shiny server and the final dropping
of GoDaddy.
}
