% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rc}
\alias{rc}
\title{render the current document}
\usage{
rc(fileName = .curFile, projName = .projName, codeDir = "GitLab",
  gitRepo = "liverpool", sysRoot = Sys.getenv("HOME"),
  userid = Sys.getenv("USER"), topDir = "public_html",
  outRoot = ".tmp", outDocxPath = outRoot, setGH = F, toPDF = F,
  toDOCX = F, toHTML = T, upload = T, livUP = FALSE, over = T,
  livPath = file.path("/media", userid, "mdrive", "public.www"),
  godPath = NULL)
}
\arguments{
\item{fileName}{file name}

\item{codeDir}{code tree path}

\item{sysRoot}{home directory}

\item{userid}{User id}

\item{topDir}{output home directory on god}

\item{outRoot}{local temporary file creation location}

\item{projDir}{project name}
}
\description{
renders the named document and puts the output html file in the correct location in my filespace
This currently assumes that I have mounted my troodon home directory to the local host via sshfs
which is available from OSXFUSE. Default paths are configured for my code locations.
}
\details{
Been making changes to this to use bookdown document styles
}
