% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rc}
\alias{rc}
\title{render the current document}
\usage{
rc(
  fileName = .curFile,
  projName = .projName,
  gitRepo = .gitRepo,
  sysRoot = Sys.getenv("HOME"),
  user = Sys.getenv("USER"),
  codeDir = file.path(sysRoot, "GitLab", gitRepo),
  hostname = "dh174037.liv.ac.uk",
  outRoot = file.path(sysRoot, ".tmp"),
  htmlRoot = file.path("/", "var", "www", "html"),
  shinyRoot = file.path("/", "u1", "shiny-server", "samples"),
  onedrive = file.path(sysRoot, "OneDrive", "me", "html"),
  docRoot = file.path(sysRoot, "Projects", projName),
  setHome = F,
  toPDF = F,
  toDOCX = F,
  toHTML = T,
  setRepo = T,
  setProj = T,
  toShiny = F,
  htmlUP = T,
  rmdUP = F,
  pdfUP = F,
  docUP = F,
  ext = "Rmd",
  dbg = F
)
}
\arguments{
\item{fileName}{file name}

\item{sysRoot}{home directory}

\item{codeDir}{code tree path}

\item{outRoot}{local temporary file creation location}

\item{projDir}{project name}

\item{userid}{User id}
}
\description{
renders the named document and puts the output html file in the correct location in my filespace
This currently assumes that I have mounted my troodon home directory to the local host via sshfs
which is available from OSXFUSE. Default paths are configured for my code locations.
}
\details{
Been making changes to this to use bookdown document styles
}
