% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rc}
\alias{rc}
\title{render the current document}
\usage{
rc(fileName = .curFile, projDir = .projName, godDir = "work",
  codeDir = "Projects", godHead = "public_html", setGH = F,
  setPH = F, toPDF = F, toDOCX = F, toHTML = T, upload = T,
  setWH = F, setPI = F, locCopy = F)
}
\arguments{
\item{fileName}{file name}

\item{projDir}{project name}

\item{codeDir}{code tree path}

\item{rootDir}{root of output tree}
}
\description{
renders the named document and puts the output html file in the correct location in my filespace
This currently assumes that I have mounted my troodon home directory to the local host via sshfs
which is available from OSXFUSE. Default paths are configured for my code locations.
}
\details{
Been making changes to this to use bookdown document styles
}
