% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcur.R
\name{rc}
\alias{rc}
\title{render the current document}
\usage{
rc(fileName = .curFile, projName = .projName, gitRepo = .gitRepo,
  sysRoot = Sys.getenv("HOME"), userid = Sys.getenv("USER"),
  codeDir = file.path(sysRoot, "GitLab", gitRepo),
  godRoot = "public_html", outRoot = file.path(sysRoot, ".tmp"),
  livRoot = file.path("/", "var", "www", "html"), setHome = F,
  toPDF = F, toDOCX = F, toHTML = T, godUP = T, livUP = FALSE,
  over = T)
}
\arguments{
\item{fileName}{file name}

\item{sysRoot}{home directory}

\item{userid}{User id}

\item{codeDir}{code tree path}

\item{outRoot}{local temporary file creation location}

\item{projDir}{project name}

\item{topDir}{output home directory on god}
}
\description{
renders the named document and puts the output html file in the correct location in my filespace
This currently assumes that I have mounted my troodon home directory to the local host via sshfs
which is available from OSXFUSE. Default paths are configured for my code locations.
}
\details{
Been making changes to this to use bookdown document styles
}
