% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmd.R
\name{rcmd}
\alias{rcmd}
\title{execute a command remotely via ssh}
\usage{
rcmd(cmd = "", crick = F, hostname = NULL, username = NULL, Xfwd = "")
}
\arguments{
\item{cmd}{the command to execute}

\item{hostname}{the remote machine}

\item{username}{currently works with ssh keys rather than password}
}
\description{
execute a command remotely via ssh
}
