% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlsFiles.R
\name{rlsFiles}
\alias{rlsFiles}
\title{list remote directory}
\usage{
rlsFiles(
  pattern,
  hostname = "localhost",
  username = "pietas",
  verb = F,
  dir = F
)
}
\arguments{
\item{pattern}{file pattern}
}
\description{
list remote directory
}
