% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFastQC.R
\name{runFastQC}
\alias{runFastQC}
\title{Submit a list of files associated with a project for  QC checking with fastQC file}
\usage{
runFastQC(projName, outRoot = file.path("/scratch/pschofield/Projects",
  projName), fileList = NULL, outName = "Analysis/fastqc",
  fileFormat = "fastq", noSub = F, scpIt = T,
  fileExt = ".fastq.gz")
}
\arguments{
\item{projName}{project name}

\item{outRoot}{output root directory}

\item{fileList}{list of fastq file}
}
\description{
Submit a list of files associated with a project for  QC checking with fastQC file
}
