% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScript.R
\name{runScript}
\alias{runScript}
\title{generate and submit a batch job}
\usage{
runScript(jname, jproj, jdesc, jscrp, db = 0,
  locroot = file.path(Sys.getenv("HOME")), logdir = NULL,
  scrpdir = NULL, remscrpdir, nnodes = 1, nproc = 8, mem = "8Gb",
  wtime = "24:00:00", xoverwrite = T, eviron = F, host = "feenix",
  jobDep = NULL)
}
\arguments{
\item{jname}{job name}

\item{jproj}{project name}

\item{jdesc}{job description}

\item{jscrp}{job script}

\item{locroot}{local root}

\item{logdir}{log directory}

\item{scrpdir}{scripts directory}

\item{nnodes}{number of nodes requested}

\item{nproc}{number of processor requested}

\item{mem}{memory size requested}

\item{wtime}{wall time requested}

\item{remroot}{remote root}

\item{overwrite}{overwrite existing script file if it exists}

\item{environ}{take environment variables of submitting process (not relevant for remote jobs)}
}
\description{
generate and submit a batch job
}
