% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcurrent.R
\name{setcurrent}
\alias{setcurrent}
\title{set current month}
\usage{
setcurrent(
  fileName,
  in_path = "/srv/http/uol/notes",
  out_path = in_path,
  linkName = "current"
)
}
\description{
set current month
}
