% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subJob.R
\name{subJob}
\alias{subJob}
\title{Submit a job to the cluster}
\usage{
subJob(scriptfile, locRoot = "/Users", remRoot = "/scratch",
  user = "pschofield", host = "troodon.scicom.picr.man.ac.uk",
  setDir = "cd /scratch/pschofield;", qsubString = "qsub", args = NULL,
  noSub = T, scpIt = F, db = F)
}
\arguments{
\item{user}{a string with the user name of the person making the ssh connection}

\item{host}{a string with the name of the login host for the cluster}

\item{args}{any arguments to add to the job submission string}

\item{noSub}{unless set to false will merely display the command rather than submit it.}

\item{scpIt}{copy to remote using scp before submit when not using FUSE}

\item{batchFile}{a string representing the name of the script file to be submitted}

\item{qsubstring}{the submission string for the torque system}
}
\description{
This functions submits a job in the form of a scriptfile to the cluster via ssh
}

