% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_yaml.R
\name{update_meta_yaml}
\alias{update_meta_yaml}
\title{update meta data}
\usage{
update_meta_yaml(
  repo_paths,
  codeDir = file.path(Sys.getenv("HOME"), "Repos"),
  db = F
)
}
\description{
update meta data
}
